Moving test_sampler_kwargs_various_methods to test_random
