+++ tests/bofire/data_models/specs/strategies.py
@@ -1,8 +1,5 @@
-import pytest
-
 import bofire.data_models.strategies.api as strategies
 import bofire.data_models.strategies.predictives.acqf_optimization
-import bofire.strategies.api as actual_strategies
 from bofire.data_models.acquisition_functions.api import (
     qEI,
     qLogNEHVI,
@@ -38,7 +35,6 @@ from bofire.data_models.strategies.api import (
     RelativeMovingReferenceValue,
     RelativeToMaxMovingReferenceValue,
 )
-from bofire.data_models.strategies.random import RandomStrategy
 from bofire.data_models.surrogates.api import BotorchSurrogates, MultiTaskGPSurrogate
 from tests.bofire.data_models.specs.api import domain
 from tests.bofire.data_models.specs.specs import Specs
@@ -1008,37 +1004,3 @@ specs.add_invalid(
     error=ValueError,
     message="Only one task can be the target fidelity",
 )
-
-
-@pytest.mark.parametrize(
-    "method,kwargs,n_samples",
-    [
-        (SamplingMethodEnum.SOBOL, {"scramble": True}, 10),
-        (SamplingMethodEnum.SOBOL, {"scramble": False}, 10),
-        (SamplingMethodEnum.LHS, {"scramble": True, "strength": 1}, 10),
-        (
-            SamplingMethodEnum.LHS,
-            {"strength": 2},
-            9,
-        ),  # strength=2 requires n to be square of prime
-        (SamplingMethodEnum.UNIFORM, {}, 10),
-    ],
-)
-def test_sampler_kwargs_various_methods(method, kwargs, n_samples):
-    """Test sampler_kwargs with various sampling methods."""
-    test_domain = Domain(
-        inputs=Inputs(
-            features=[
-                ContinuousInput(key="x1", bounds=(0, 1)),
-                ContinuousInput(key="x2", bounds=(0, 2)),
-                ContinuousInput(key="x3", bounds=(0, 3)),
-            ]
-        ),
-        outputs=Outputs(features=[ContinuousOutput(key="y")]),
-    )
-    data_model = RandomStrategy(
-        domain=test_domain, fallback_sampling_method=method, sampler_kwargs=kwargs
-    )
-    strategy = actual_strategies.map(data_model=data_model)
-    candidates = strategy.ask(n_samples)
-    assert len(candidates) == n_samples
